+++ apps/openmw/mwphysics/trace.cpp
@@ -67,10 +67,10 @@ namespace MWPhysics
         // Copy the hit data over to our trace results struct:
         if (traceCallback.hasHit())
         {
-            mFraction = traceCallback.m_closestHitFraction;
+            mFraction = static_cast<float>(traceCallback.m_closestHitFraction);
             // ensure fraction is correct (covers intended distance traveled instead of actual distance traveled)
             if (doingShortTrace && (end - start).length2() > 0.0)
-                mFraction *= (btend - btstart).length() / (end - start).length();
+                mFraction *= static_cast<float>((btend - btstart).length() / (end - start).length());
             mPlaneNormal = Misc::Convert::toOsg(traceCallback.m_hitNormalWorld);
             mEndPos = (end - start) * mFraction + start;
             mHitPoint = Misc::Convert::toOsg(traceCallback.m_hitPointWorld);
@@ -85,7 +85,7 @@ namespace MWPhysics
 
                 if (newTraceCallback.hasHit())
                 {
-                    mFraction = newTraceCallback.m_closestHitFraction;
+                    mFraction = static_cast<float>(newTraceCallback.m_closestHitFraction);
                     mPlaneNormal = Misc::Convert::toOsg(newTraceCallback.m_hitNormalWorld);
                     mEndPos = (end - start) * mFraction + start;
                     mHitPoint = Misc::Convert::toOsg(newTraceCallback.m_hitPointWorld);
@@ -109,7 +109,7 @@ namespace MWPhysics
             actor->getCollisionObject(), Misc::Convert::toBullet(start), Misc::Convert::toBullet(end), world, true);
         if (traceCallback.hasHit())
         {
-            mFraction = traceCallback.m_closestHitFraction;
+            mFraction = static_cast<float>(traceCallback.m_closestHitFraction);
             mPlaneNormal = Misc::Convert::toOsg(traceCallback.m_hitNormalWorld);
             mEndPos = (end - start) * mFraction + start;
         }
